requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
pydantic
pydantic[email]
datetime
loguru == 0.7.2
pytest
mypy
python-dotenv==1.0.0
httpx
alembic
asyncpg==0.29.0
psycopg2
pytest-asyncio
trio
cryptography==41.0.2
sqlalchemy
